# pos_app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: POS Multi-Sucursal API
  version: 1.0.0
  description: Punto de venta y control de inventario multi-sucursal.
paths:
...
  /api/v1/sales/:
    get:
      operationId: sales_list
      description: Ventas de la sucursal, de la más reciente a la más antigua.
      parameters:
      - in: query
        name: branch
        schema:
          type: string
          maxLength: 20
          minLength: 1
      - in: query
        name: cash_session
        schema:
          type: integer
          minimum: 1
      - in: query
        name: date_from
        schema:
          type: string
          format: date-time
      - in: query
        name: date_to
        schema:
          type: string
          format: date-time
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      tags:
      - sales
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSaleList'
          description: ''
    post:
      operationId: sales_create
      description: |-
        Registra una venta en la caja abierta del usuario.

        Los precios se toman de la base de datos y la tasa activa queda
        congelada en la venta.
      tags:
      - sales
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSale'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateSale'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateSale'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
          description: ''
  /api/v1/sales/{id}/:
    get:
      operationId: sales_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - sales
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
          description: ''
  /api/v1/sales/reports/summary/:
    get:
      operationId: sales_reports_summary
      description: Número de ventas y totales en USD y VES del periodo.
      parameters:
      - in: query
        name: branch
        schema:
          type: string
          maxLength: 20
          minLength: 1
      - in: query
        name: cash_session
        schema:
          type: integer
          minimum: 1
      - in: query
        name: date_from
        schema:
          type: string
          format: date-time
      - in: query
        name: date_to
        schema:
          type: string
          format: date-time
      tags:
      - sales
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesSummary'
          description: ''
...
components:
  schemas:
...
    CreateSale:
      type: object
      properties:
        branch:
          type: string
          maxLength: 20
        customer_tax_id:
          type: string
          maxLength: 20
        customer_name:
          type: string
          maxLength: 150
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItemInput'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentInput'
      required:
      - items
      - payments
    CurrencyEnum:
      enum:
      - VES
      - USD
      type: string
      description: |-
        * `VES` - Bolívares
        * `USD` - Dólares
...
    MethodEnum:
      enum:
      - POS_CARD
      - CASH_VES
      - CASH_USD
      - MOBILE_PAYMENT
      type: string
      description: |-
        * `POS_CARD` - Punto de venta
        * `CASH_VES` - Efectivo VES
        * `CASH_USD` - Efectivo USD
        * `MOBILE_PAYMENT` - Pago móvil
...
    PaginatedSaleList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Sale'
...
    PaymentInput:
      type: object
      properties:
        method:
          $ref: '#/components/schemas/MethodEnum'
        currency:
          $ref: '#/components/schemas/CurrencyEnum'
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        approval_reference:
          type: string
          maxLength: 50
      required:
      - amount
      - currency
      - method
...
    Sale:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        cash_session:
          type: integer
          readOnly: true
        user:
          type: integer
          readOnly: true
        branch:
          type: string
          readOnly: true
        customer_tax_id:
          type: string
          readOnly: true
        customer_name:
          type: string
          readOnly: true
        exchange_rate_at_invoice:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,4})?$
          readOnly: true
        total_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        total_ves:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        details:
          type: array
          items:
            $ref: '#/components/schemas/SaleDetail'
          readOnly: true
        payments:
          type: array
          items:
            $ref: '#/components/schemas/SalePayment'
          readOnly: true
      required:
      - branch
      - cash_session
      - created_at
      - customer_name
      - customer_tax_id
      - details
      - exchange_rate_at_invoice
      - id
      - payments
      - total_usd
      - total_ves
      - user
    SaleDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        product:
          type: integer
          readOnly: true
        quantity:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        unit_price_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        subtotal_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - id
      - product
      - quantity
      - subtotal_usd
      - unit_price_usd
    SaleItemInput:
      type: object
      description: 'El cliente nunca envía el precio: se toma de la base de datos.'
      properties:
        product_id:
          type: integer
          minimum: 1
        quantity:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
      required:
      - product_id
      - quantity
    SalePayment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        method:
          allOf:
          - $ref: '#/components/schemas/MethodEnum'
          readOnly: true
        currency:
          allOf:
          - $ref: '#/components/schemas/CurrencyEnum'
          readOnly: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        approval_reference:
          type: string
          readOnly: true
      required:
      - amount
      - approval_reference
      - currency
      - id
      - method
    SalesSummary:
      type: object
      properties:
        sales_count:
          type: integer
        total_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        total_ves:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
      required:
      - sales_count
      - total_usd
      - total_ves
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT